pyproject.toml
[tool.poetry]
name = "converters"
version = "1.0.0"
description = "Write conversions between python objects using an expressive and declarative class-based API"
authors = ["7bridges <os@the7bridges.com>"]

[tool.poetry.dependencies]
python = ">=3.6.1"
lxml = {version = ">=4.6.3", optional = true}
pydash = ">=4.9.3"
python-dateutil = {version = ">=2.8.1", optional = true}
pytz = {version = ">=2021.1", optional = true}
Django = {version = ">=2.2", optional = true}
Unidecode = {version = ">=1.2.0", optional = true}
pandas = {version = ">=0.25.3", optional = true}
django-countries = {version = ">=7.1", optional = true}

[tool.poetry.dev-dependencies]
pytest = ">=6.2.2"
syrupy = ">=1.2.2"
pytest-django = "^4.2.0"
pytest-mock = "^3.6.0"

[tool.poetry.extras]
# ℹ️ if you update any of these, be sure to update "all = [...]" as well, so that it's easy to get
# all dependencies for testing
datetime = ["python-dateutil", "pytz"]
django = ["Django", "django-countries"]
django_xml = ["Django", "django-countries", "unidecode", "lxml"]
pandas = ["pandas"]
xml = ["lxml"]

all = ["Django", "django-countries", "lxml", "pandas", "python-dateutil", "pytz", "unidecode"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
